requirements.txt
piper-tts
yt_dlp
openwakeword
pyaudio
faster-whisper
google-generativeai